requirements.txt
dash==2.14.2
jupyter-dash==0.4.2
dash_daq==0.5.0
dash-bootstrap-components==1.5.0
dash-table==5.0.0
gspread==6.0.0
oauth2client==4.1.3
easydict==1.11
pandas==1.3.0
pyyaml